latest commit on part 1
