+++ .eslintrc
@@ -1,25 +1,12 @@
 {
+    "extends": "airbnb",
     "env": {
-        "es6": true
-    },
+      "jest": true,
+      "node": true,
+      "browser": true
+    },  
     "parser": "babel-eslint",
-    "plugins": [
-        "react"
-    ],
-    "parserOptions": {
-        "ecmaVersion": 6,
-        "sourceType": "module",
-        "ecmaFeatures": {
-            "experimentalObjectRestSpread": true,
-            "jsx": true
-        }
-    },
-    "rules": {   
-        "react/jsx-uses-react": "error",   
-        "react/jsx-uses-vars": "error",
-        "semi": [ 2, "always" ],
-        "quotes": [ 2, "single" ],
-        "indent": [ 2, 4 ],
-        "no-unused-vars": "error"
+    "rules": {
+      "import/no-extraneous-dependencies": ["error", { "devDependencies": ["/*", "**/*.config.js", "**/*.test.js", "**/*.spec.jsx"]}]
     }
-}
+  }
