+++ config/jest/jestsetup.js
@@ -1,19 +1,19 @@
 module.exports = {
-    rootDir: '../../',
-    clearMocks: true,
-    moduleNameMapper: {
-        '\\.(css|sass|scss)$': 'identity-obj-proxy',
+  rootDir: '../../',
+  clearMocks: true,
+  moduleNameMapper: {
+    '\\.(css|sass|scss)$': 'identity-obj-proxy',
+  },
+  resolver: null,
+  collectCoverage: true,
+  collectCoverageFrom: ['src/**/*.{js,jsx}'],
+  coveragePathIgnorePatterns: ['src/index.js'],
+  coverageThreshold: {
+    global: {
+      branches: 100,
+      functions: 100,
+      lines: 100,
+      statements: 100,
     },
-    resolver: null,
-    collectCoverage: true,
-    collectCoverageFrom: ['src/**/*.{js,jsx}'],
-    coveragePathIgnorePatterns: ['src/index.js'],
-    coverageThreshold: {
-        global: {
-            branches: 100,
-            functions: 100,
-            lines: 100,
-            statements: 100,
-        },
-    },
+  },
 };
