+++ config/webpack.production.config.js
@@ -3,63 +3,69 @@ const path = require('path');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
 const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+const CleanWebpackPlugin = require('clean-webpack-plugin');
 const webpack = require('webpack');
 
 module.exports = merge(require('./webpack.base.config'), {
-    output: {
-        path: path.resolve(__dirname, '../build'),
-        filename: '[name].[hash].js',
-        publicPath: 'build',
-    },
-    mode: 'production',
-    devtool: 'none',
-    entry: [
-        path.join(__dirname, '../src/index.js'),
-    ],
-    optimization: {
-        runtimeChunk: 'single',
-        splitChunks: {
-            cacheGroups: {
-                vendor: {
-                    test: /[\\/]node_modules[\\/]/,
-                    name: 'vendors',
-                    chunks: 'all',
-                },
-            },
+  output: {
+    path: path.resolve(__dirname, '../build'),
+    filename: '[name].[hash].js',
+    publicPath: 'build',
+  },
+  mode: 'production',
+  devtool: 'none',
+  entry: [
+    path.join(__dirname, '../src/index.jsx'),
+  ],
+  optimization: {
+    runtimeChunk: 'single',
+    splitChunks: {
+      cacheGroups: {
+        vendor: {
+          test: /[\\/]node_modules[\\/]/,
+          name: 'vendors',
+          chunks: 'all',
         },
-    
-        minimizer: [
-            new UglifyJsPlugin({
-                cache: true,
-                parallel: true,
-                sourceMap: true,
-            }),
-            new OptimizeCSSAssetsPlugin({}),
-        ],
+      },
     },
-    module: {
-        rules: [
-            {
-                test: /\.(sass|scss)$/,
-                exclude: /node_modules/,
-                use: [
-                    MiniCssExtractPlugin.loader,
-                    {
-                        loader: 'css-loader',
-                        options: {
-                            modules: true,
-                            localIdentName: '[name]__[local]__[hash:base64:5]',
-                        },
-                    },
-                    'sass-loader',
-                ],
-            },
-        ],
-    },
-    plugins: [
-        new MiniCssExtractPlugin({
-            filename: 'css/[name].[contenthash].css',
-        }),
-        new webpack.HashedModuleIdsPlugin(),
+
+    minimizer: [
+      new UglifyJsPlugin({
+        cache: true,
+        parallel: true,
+        sourceMap: true,
+      }),
+      new OptimizeCSSAssetsPlugin({}),
     ],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.(sass|scss)$/,
+        exclude: /node_modules/,
+        use: [
+          MiniCssExtractPlugin.loader,
+          {
+            loader: 'css-loader',
+            options: {
+              modules: true,
+              localIdentName: '[name]__[local]__[hash:base64:5]',
+            },
+          },
+          'sass-loader',
+        ],
+      },
+    ],
+  },
+  plugins: [
+    new MiniCssExtractPlugin({
+      filename: 'css/[name].[contenthash].css',
+    }),
+    new webpack.HashedModuleIdsPlugin(),
+    new CleanWebpackPlugin(['build'], {
+      root: process.cwd(),
+      verbose: true,
+      dry: false,
+    }),
+  ],
 });
