+++ src/components/App/App.jsx
@@ -1,16 +1,11 @@
 import React from 'react';
 import './App.scss';
-import Footer from '../Footer/Footer.jsx';
+import Footer from '../Footer/Footer';
 
+const App = () => (
+  <div className="main_container">
+    <Footer />
+  </div>
+);
 
-class App extends React.Component {
-    render() {
-        return(
-            <div className="main_container">
-                <Footer />
-            </div>
-        );
-    }
-};
-
-export default App; 
+export default App;
