+++ src/components/App/index.jsx
@@ -1,1 +1,1 @@
-export {default} from './App.jsx';
+export { default } from './App';
