+++ src/components/Footer/index.jsx
@@ -1,1 +1,1 @@
-export {default} from './Footer.jsx';
+export { default } from './Footer';
