+++ src/components/Root.jsx
@@ -1,4 +1,4 @@
 import { hot } from 'react-hot-loader';
-import App from './App/App.jsx';
+import App from './App/App';
 
 export default hot(module)(App);
