--- src/components/__tests__/Root.spec.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-
-import Root from '../Root.jsx';
-
-describe('Root', () => {
-    it('Root renders correctly', () => {
-        const renderedRoot = renderer.create(<Root />).toJSON();
-        expect(renderedRoot).toMatchSnapshot();
-    });
-});
