+++ src/index.html
@@ -4,6 +4,5 @@
     </head>
     <body>
         <div id="app"></div>
-        <script src="build/main.js"></script>
     </body>
 </html>
