+++ src/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Root from './components/Root.jsx';
+import Root from './components/Root';
 
 ReactDOM.render(<Root />, document.querySelector('#app'));
